w_ZFAIL fix

+++ py/LSS/ssr_tools_new.py
@@ -172,18 +172,20 @@ class model_ssr_zfac:
         self.maxz = maxz
         
         self.res_mod_slp = self.get_slpfunc()
-        fo = open(outdir+outfn_root+'slp_wzfac.txt','w')
-        fo.write(str(self.res_mod_slp))
-        fo = open(outdir+outfn_root+'slp_vszfac.txt','w')
-        for i in range(0,len(self.slpl)):
-            fo.write(str(self.zfacl[i])+' '+str(self.slpl[i])+'\n')
-        fo.close()
 
+        if not os.path.exists(outdir+outfn_root+'slp_wzfac.txt'):
+            fo = open(outdir+outfn_root+'slp_wzfac.txt','w')
+            fo.write(str(self.res_mod_slp))
+            fo.close()
+        if not os.path.exists(outdir+outfn_root+'slp_vszfac.txt')
+            fo = open(outdir+outfn_root+'slp_vszfac.txt','w')
+            for i in range(0,len(self.slpl)):
+                fo.write(str(self.zfacl[i])+' '+str(self.slpl[i])+'\n')
+            fo.close()
         #fo.write('#a b c chi2\n')
         #for par in pars:
         #   fo.write(str(par)+' ')
         #fo.write(str(chi2)+'\n')
-        fo.close()
 
         self.ssrtot = len(self.cat[self.selgz])/len(self.cat)
         self.tracer = tracer
